requirements.txt
aniso8601==10.0.1
attrs==25.4.0
blinker==1.9.0
click==8.1.8
Flask==3.1.2
flask-restx==1.3.2
importlib_metadata==8.7.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
referencing==0.36.2
rpds-py==0.27.1
typing_extensions==4.15.0
Werkzeug==3.1.5
zipp==3.23.0
